+++ srcs/push_array_into_stack.c
@@ -6,13 +6,13 @@
 /*   By: pcunha <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/17 01:48:42 by pcunha            #+#    #+#             */
-/*   Updated: 2021/07/18 18:05:49 by pcunha           ###   ########.fr       */
+/*   Updated: 2021/07/20 18:22:33 by pcunha           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-t_dlist	*push_array_into_stack(int *array, int n_array, t_dlist *list)
+t_dlist	*push_array_into_stack(double *array, int n_array, t_dlist *list)
 {
 	int i;
